+++ public/models/contribution-month/test/contribution-month_test.js
@@ -11,7 +11,13 @@ import MonthlyOSProject from '../monthly-os-project';
 import MonthlyClientProject from '../monthly-client-project';
 import fixture from 'can-fixture';
 
-QUnit.module( 'models/contribution-month' );
+QUnit.module( 'models/contribution-month', {
+	setup: function(){
+		localStorage.clear();
+		// Reset fixture store before every test:
+		store.reset();
+	}
+});
 
 QUnit.test( "ContributionMonth basic test", function() {
 
@@ -291,4 +297,53 @@ QUnit.test( "OS Projects and Client Projects are provide sorted clones", functio
   QUnit.equal(sortedProjects[1].osProjectRef.value.name, 'C');
   QUnit.equal(sortedClients[0].clientProjectRef.value.name, 'X');
   QUnit.equal(sortedClients[1].clientProjectRef.value.name, 'Z');
+});
+
+QUnit.test("getList of ContributionMonth", function(assert) {
+	let done = assert.async();
+	ContributionMonth.getList({}).then(function(contributionMonths) {
+
+		// TODO: check if we need to test against `clientProjectRef.value`.
+		QUnit.ok(contributionMonths[0].monthlyClientProjects[0].clientProjectRef._id === "1-Levis", 'contains a client project');
+		var first = contributionMonths[0].monthlyOSProjects[0].osProjectRef._id,
+			second = contributionMonths[0].monthlyClientProjects[0].monthlyClientProjectsOSProjects[0]._id;
+
+		QUnit.ok(first, 'first exists');
+		QUnit.ok(first === second, 'first and second are equal');
+
+		done();
+	}, function(err) {
+		QUnit.ok(false, err);
+	});
+});
+
+QUnit.test("make type convert able to accept instances (#23)", function() {
+	var osProject = new OSProject({
+		_id: "somethingCrazey",
+		name: "CanJS"
+	});
+
+	var clientProject = new ClientProject({
+		_id: "asl;dfal;sfj ;lakwj",
+		name: "HualHound"
+	});
+
+	var contributionMonth = new ContributionMonth({
+		_id: "aslkfalsjklas",
+		date: 124234211310000,
+		monthlyOSProjects: [{
+			significance: 80,
+			commissioned: true,
+			osProjectRef: osProject._id,
+			osProject: osProject
+		}],
+		monthlyClientProjects: [{
+			monthlyClientProjectsOSProjects: [ osProject ],
+			hours: 100,
+			clientProjectRef: clientProject._id,
+			clientProject: clientProject
+		}]
+	});
+
+	QUnit.equal(contributionMonth.monthlyOSProjects[0].osProject.name , "CanJS");
 });
